APM: Wrong device selection for music during concurrency.
Music playback is switching from WFD to speaker when ringtone is
played and speaker to hanset when ringtone is played.

The issue is wrong device is getting returned from getNewOutputDevices
which selects device on basis of active product strategy so wrong device is
selected based on the order of product strategies.

To avoid this we add a check to see if the Alarm stream or Enforced Audible
stream is actually active,if current output desc has VOICE_CALL stream.

CRs-Fixed: 2446622
Change-Id: Iff6aa67f799cec5f960707fb39333393944d1478
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: ghou8s <ghou8s@gmail.com>